feat：流水线版本管理机制 #8161 补充task的env
